+++ app/src/main/java/com/silverknife/meizhi/TabPresenter.kt
@@ -1,0 +1,9 @@
+package com.silverknife.meizhi
+
+import com.silvericekey.skutilslibrary.base.BasePresenter
+
+class TabPresenter:BasePresenter() {
+    override fun onDestroy() {
+
+    }
+}