+++ app/src/main/java/com/silverknife/meizhi/TabsPresenter.kt
@@ -1,0 +1,10 @@
+package com.silverknife.meizhi
+
+import com.silvericekey.skutilslibrary.base.BasePresenter
+
+class TabsPresenter: BasePresenter() {
+    override fun onDestroy() {
+
+    }
+
+}