+++ gluetablayout/src/main/java/com/negier/gluetablayout/BasePagerAdapter.java
@@ -1,0 +1,14 @@
+package com.negier.gluetablayout;
+
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentStatePagerAdapter;
+
+public abstract class BasePagerAdapter extends FragmentStatePagerAdapter {
+    public BasePagerAdapter(FragmentManager fm) {
+        super(fm);
+    }
+
+    public int getPageIcon(int position) {
+        return 0;
+    }
+}